Update patch set 4
Patch Set 4:

FWIW, I'm having third thoughts about this patch.

The RADOS backend stores a binary blob already. We convert that to a text string, which is semi-silly. Also, it's technically possible to embed NULLs in the long-form client id and we don't encode those properly today.

We probably ought to change the recovery backends to track this as a blob and length instead of in strings. I think I'm going to rework this code along those lines. It's a larger patchset, but it's a better design, and we don't want to have to make this change more than once.

Patch-set: 4
Reviewer: Gerrit User 1008909 <1008909@d5d70762-12d0-45a1-890d-524b12d3f735>