Fix the mounted_on_fileid attribute for referral directories
When returning attributes for the referral directories, mounted_on_fileid
attribute was set incorrectly in getattr. It was always set to zero.
Because of this all the referrals were treated as the same directory by the
NFS client. So, whichever referral is mounted first, all other referrals
will also point to the same export even though the fs locations were set
correctly.

Change-Id: I259ae04e8ea542682abd19fa37a1e70edef9856d
Signed-off-by: Sriram Patil <sriramp@vmware.com>