Update patch set 1
Patch Set 1:

That really seems like a bug in the spec, tbh. The CB_NULL was originally done for v4.0, since the backchannel was a separate socket (and initiated by the server). FWIW, knfsd stopped doing this in 2015 -- see kernel  commit 4bd9e9b77fc6787c45b8bb439f6511aa3478606c.

My understanding is this: With v4.1+ and sessions, we always associate the backchannel with a client-originated connection and in the context of a CREATE_SESSION or BIND_CONN_TO_SESSION request from the client. If the socket isn't working, then the client will never see the reply to those calls, so testing the connection is not terribly valuable at that point.

I suppose we could argue for doing this in order to verify that the backchannel RPC program is up and accepting requests, but I'm not sure that's really worthwhile.

Patch-set: 1
Reviewer: Gerrit User 1008909 <1008909@d5d70762-12d0-45a1-890d-524b12d3f735>