Update patch set 1
Patch Set 1:

I think this is fine.  My thinking on this had been that potentially, the decision to rely on refs of drc to control transition from active to recycle_q, but this potentially more consistent with a model where trunked connections share a drc--that probably can't happen today (and isn't spec nfs), but Dan mentioned it, and I think for v3|4.0, sharing the drc would be correct.

Patch-set: 1
Reviewer: Gerrit User 1004075 <1004075@d5d70762-12d0-45a1-890d-524b12d3f735>