Fix NLM owner refcount leak in state_lock()
We copy conflict for blocking and non-blocking locks in state_lock(). We
do send STATE_LOCK_CONFLICT for non-blocking calls where the caller
releases the hold on the owner. We send STATE_LOCK_BLOCKED for blocked
locks, so we need to release the hold in state_lock() itself.

copy conflict only for non-blocking NLM requests.

Change-Id: I7eb7c7a72d9f748f97573327ef6f17406a16b494
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>