Update patch set 1
Patch Set 1:

I'm not fully up on the NLM codepaths.  This is done in nlm4_Test_Free(), and that seems to be the right place to do it.  Why is a Test op being freed in nlm4_Lock_Free()?

Patch-set: 1
Reviewer: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>