Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review+1
> > 
> > This all looks fine to me. Do we need to do anything differently in the RPM specfile with this?
> 
> No, nothing is needed for RPMs or DEBs.  It's all self contained in the .so

The $10 question is: is this all the APIs that a FSAL writer needs.  I did this empirically by linking FSALs and ganesha.nfsd (and the other apps) to discover which functions they call in the library. Just because nothing currently currently calls a particular function doesn't necessarily or automatically mean it shouldn't be exported from the .so.

Patch-set: 2
Reviewer: Gerrit User 1004389 <1004389@d5d70762-12d0-45a1-890d-524b12d3f735>