Update patch set 2
Patch Set 2:

> > The $10 question is: is this all the APIs that a FSAL writer
 > needs.  I did this empirically by linking FSALs and ganesha.nfsd
 > (and the other apps) to discover which functions they call in the
 > library. Just because nothing currently currently calls a
 > particular function doesn't necessarily or automatically mean it
 > shouldn't be exported from the .so.
 > 
 > I think this is a fine place to start. It's always easier to
 > whitelist symbols after the fact than to blacklist them.

I agree. As long as all the in-tree FSALs work, we should be good. If an out of tree FSAL is using an API we don't expect to be used, then hopefully we'll hear from them and we can either regularize the API or we can tell them why they should not rely on that API.

Patch-set: 2
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>