Update patch set 1
Patch Set 1:

So, this looks technically correct.  However, I wouldn't just add a ton of tracepoints from the start, especially if you don't know what info is needed for each one.  Rather, add a few sample tracepoints, to get the pattern in place, and then add more as they are needed for debugging. Just listing the functions called is generally not very useful trace output.

Tracepoints are especially useful for refcounts and making sure things that are created are released.  I don't think FSAL Gluster has any refcounts, so my suggestion is that you add tracepoints for FDs.  Trace when FDs are created and released, and possibly a bit up the call stack if that seems to be necessary (for example if you need to associate an FD with a state_t).

Patch-set: 1
CC: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>