Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1:
> > 
> > So, this looks technically correct.  However, I wouldn't just add a ton of tracepoints from the start, especially if you don't know what info is needed for each one.  Rather, add a few sample tracepoints, to get the pattern in place, and then add more as they are needed for debugging. Just listing the functions called is generally not very useful trace output.
> > 
> > Tracepoints are especially useful for refcounts and making sure things that are created are released.  I don't think FSAL Gluster has any refcounts, so my suggestion is that you add tracepoints for FDs.  Trace when FDs are created and released, and possibly a bit up the call stack if that seems to be necessary (for example if you need to associate an FD with a state_t).
> 
> Ok yeah that makes sense, and what exactly did you mean by up the call? I didn't understand this bit properly.

By "up the stack", I mean further up the call stack.  So if a() calls b(), which calls c(), which creates an FD, then you trace in c(), but maybe you also want to trace in b().  This can be done either by creating a new tracepoint and using it in b(), or (if c() isn't useful itself) modifying c() to take a function/line pair, and passing in the name of the caller, and tracing that.  See mdcache_lru_unref() for an example of this second one.  It uses a macro to pass in __func__ and __LINE__, and then uses that in it's tracepoint.  This is because all refs are dropped in this function, so tracing it's name isn't very useful.

Patch-set: 2