Update patch set 1
Patch Set 1:

So, I'm not sure this is right.  If lookup() fails, it means this is the top of the filesystem.  In that case, the parent pointer should point into a different FSAL (FSAL_PSEUDO, at least), and so shouldn't be subject to the expire_parent_time setting of this FSAL.  I think we may need to mark the top of the FS somehow, and not expire it's parent pointer, or that pointer can never be refreshed, and we can never call lookupp() on it.

In other words, fsdir.parent belongs to the FSAL owning parent handle, not the FSAL owning the current handle, so it shouldn't be subject to the current FSAL's timeout.

Am I missing something?

Patch-set: 1
CC: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>