Free the extra ref taken by the dirent on the mdcache entry even if the content lock is held in read mode.
Not freeing the extra ref leads to memory growth as we are not able to prune
the mdcache entries with ref count >= 2.
Though we expect that the ref will be released on a invalidate call,
but invalidation of the dirents only happen when we either add/delete any directory entry.
Otherwise the dirent stays in the cache pinning the correspodning mdcache entry.

Signed-off-by: Ashish Sangwan <ashishsangwan2@gmail.com>
Change-Id: I48c6e6f02705a067e3ce0fcaa2177e069ddd53f1