Update patch set 3
Patch Set 3: Code-Review-1

This won't work.  The whole point of dirent->entry is to ensure the entry stays in the cache until after the callback is called, so we cannot drop the ref here, we must wait until after the callback is called.  This will result in the entry being freed here, and re-created below.

What the code is currently doing is taking a local ref for this thread to keep the entry in cache until after the callback is called, and then dropping it afterwards.  The ref on dirent->entry is only dropped in the thread with the write lock (which is the one that created that ref).  So it should be safe now.

If there is an aborted readdir, then some of the dirents (ones in the current chunk that were not processed) will stay ref'd until the directory is invalidated.  This is okay, it's a small number of dirents.

Is there an actual leak here?  Or is the problem just the aborted readdir leaving ref's for a while?

Patch-set: 3
Reviewer: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1