Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> This won't work.  The whole point of dirent->entry is to ensure the entry stays in the cache until after the callback is called, so we cannot drop the ref here, we must wait until after the callback is called.  This will result in the entry being freed here, and re-created below.
> 


What if we reverse the order?
First we call mdcache_find_keyed_reason on the dirent's key, that will grab the ref and make sure the entry stays in the cache. Next we check if dirent->entry is valid and atomically sets the entry pointer to NULL and drop the ref.

> What the code is currently doing is taking a local ref for this thread to keep the entry in cache until after the callback is called, and then dropping it afterwards.  The ref on dirent->entry is only dropped in the thread with the write lock (which is the one that created that ref).  So it should be safe now.
> 
> If there is an aborted readdir, then some of the dirents (ones in the current chunk that were not processed) will stay ref'd until the directory is invalidated.  This is okay, it's a small number of dirents.
> 
> Is there an actual leak here?  Or is the problem just the aborted readdir leaving ref's for a while?


It's not about aborted readdir. For example, we populated 128 entries while populating a chunk under the content lock held in write mode and we are only able to send back to the client 64 entries in that readdir RPC. In the next readdir iteration content lock will be held in read mode and we won't be able to drop the refs for other 64 entries, unless the dirent cache invalidated. These entries with an extra ref are occupying the LRU and makes the reaping impossible. On our system, we are reaching in millions of entries in LRU without the patch. After the patch entries remains equal to the watermark set (100k)

Patch-set: 3
Reviewer: Gerrit User 1016062 <1016062@d5d70762-12d0-45a1-890d-524b12d3f735>