Use display_buffer for logging file handles, add new Log macros
New Log macros are introduced to make it easier to log messages
that include file handles. Particularly for NFS v3 operations,
LogNFS3_Operation logs the entry to an operation complete with
handle and any additional parameters desired. LogNFS3_Operation2
is for the operations with two handles.

LogHandleNFS4 replaces sprint_fhandle4.

LogCompoundFH is turned into a macro for better line number
reporting.

All the sprint_handle functions are removed as well as the now
uneccesary sprint_mem.

Change-Id: Ie76757a5115a4a541dcee72f6e62d20ed61f2577
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>