coverity: string is not null terminated
Originally intended as a test of CI failures, this fixes an issue (of
sorts) reported by coverity, namely not enough memory is allocated
to store a terminating NULL.

The warning is a false positive. utf8strings in ganesha are
ostensibly strings that come off the wire and may or may not be null
terminated; given that, it is not safe to call strlen(). Furthermore,
they need not be NULL terminated when stored internally as the length
is stored in an associated length variable.

In this case the string comes from an xattr and should contain a ':'
separator. The ':' separator, if there is one, is replaced with a null
by strsep(). By convention it should still be considered not safe to
call strlen() or (gsh_)strdup().

The string length is now calculated, semi-opportunistically allowing
for a terminating NULL and copying the NULL. Although the terminating
NULL is not necessary, this change does silence the coverity warning.

Unfortunately simply adding a /* coverity[OVERRUN] */ comment here
did not silence it.

Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>
Change-Id: I594ce69f362d2a0db9a5e9257dc6b032a6552791