coverity: string is not null terminated
Now with utf8string_dup().

Fixes an issue (of sorts) reported by coverity, namely not enough
memory is allocated to store a terminating NULL in a utf8string.

The warning is a false positive. utf8strings in ganesha are nominally
strings that come off the wire and may or may not be null terminated;
given that, it is not safe to call strlen(). Furthermore, they need
not be NULL terminated when stored internally as the length is stored
in an associated length variable.

In this case the string comes from an xattr and should contain a ':'
separator. The ':' separator, if there is one, is replaced with a null
by strsep(). By convention it should still be considered not safe to
call strlen() or (gsh_)strdup(),

Unfortunately simply adding a /* coverity[OVERRUN] */ comment here
did not silence it.

Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>
Change-Id: I3f1b3840022686d979421d267b6fc518e01fdd02