Update patch set 21
Patch Set 21:

Not sure why I got an email for this old patch few days ago! In any case, this needs to be re-visited to really support mixed endian types as Linux NFS client (and some other NFS clients too) doesn't handle multiple file handles to the same file/object. This feature can't be used the way it is implemented now.

We should stick to one unique file handle (big endian or small endian) for an object and deal with it.

Patch-set: 21