Update patch set 1
Patch Set 1:

I don't believe this should be possible to hit.  The only caller is mdcache_new_entry(), which first calls cih_get_by_key_latch(), which takes a write lock on the partition.  If the entry was in the tree, it will be found by that.  If it wasn't, no one can add it, because we have the lock, so it won't be there when we set it.

Has this actually been hit?  Or is this a case of code inspection?  If it's been hit, something is badly wrong with our locking on the hash table, I suspect.

Patch-set: 1
Reviewer: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>