Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This is based on code inspection.

Okay, so it should be safe in current use; but if you want to handle this case, then cih_set_latched() should be modified to return a pointer, as described in it's comment, and it should return the existing entry if there is one.  Then, in mdcache_new_entry(), you can treat this the same as if cih_get_by_key_latch() returned an existing entry.

An alternative may be to just add a comment at the avltree_insert() line that the latch protects against a race inserting it, and it's safe to ignore the return.

Patch-set: 1
Reviewer: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>