Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

The issue with tabs in the python file needs to be addressed.

Also, it looks like the python code computes a duration using a start time from Ganesha and a current time computed locally. Is it ever possible that the dBus client and this python code are running on a different host?

Patch-set: 1
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1