Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Ok, that seems reasonable, though there is something to be said for the number missed to be in the same message. I'd like to hear your thoughts on what messages you would like rate limited.
> 
> Good question! A perfect message would be "stalled" message when a connection exceeds Dispatch_Max_Reqs_Xprt reqs. I thought about idmapper failures but I think they should be debug messages, right now, I put them at INFO level in a patch I posted.
> 
> Some examples would be:
> 
> LogCrit(COMPONENT_NFS_V4, "Bad Minor Version %d"
> LogInfo(COMPONENT_NFS_V4, "Unsupported minor version %d",
> "MOUNT: Performance warning: Export entry is not cached"

Ok, yea, those would be good candidates.

I would still prefer to see the missed count in the same message as the main log message (so when filtering you will see it). It actually could be a new column in the format, with config to show it or not... (then it could be a simple notation : x1 : or : x1000 :...)

Patch-set: 4
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>