Coverity: dead code in fsal_create
A "if" condition checks at the beginning to see whether the enumeration 'type'
is 1~7(REGULAR_FILE~DIRECTORY). In the switch statement, the NO_FILE_TYPE is 0,
and EXTENDED_ATTR is 8. So the case NO_FILE_TYPE and EXTENDED_ATTR is dead code.

The "if" condition at the beginning should be removed.

Change-Id: If8b2153072134dc5738fd56980bbacf6533336cd
Signed-off-by: Shangzhong Zhu <zhu.shangzhong@zte.com.cn>