mdcache_create_handle: Don't lookup for parent for a root object
Currently in mdcache_create_handle() after getting an 'entry' we
call mdc_get_parent() for the 'entry'. But if the entry->obj_handle
points to root object then we would be doing lookup of parent for the
root, which we should not be doing.
To avoid this prior to call mdc_get_parent() check if the
entry->obj_handle points to a root object, if yes then don't call
mdc_get_parent() for the 'entry'.

Change-Id: Ibf25fd283573c9fa28ae0452d96f00f294f66828
Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>