Coverity: check if data->session is NULL after deref
In nfs4_op_reclaim_complete, the pointer data->session has been
dereferenced in the statement "nfs_client_id_t
*clientid = data->session->clientid_record;" at the beginning. The "if"
condition "if (data->session == NULL)" is no effect. So the "if"
condition may be removed.

The "if" condition "if (data->minorversion == 0)" is also redundant. We
check for version op-overflow in nfs4_Compound.c.

Change-Id: Ia3d62cc8957aca9e006fa954619126639cdda9fa
Signed-off-by: Shangzhong Zhu <zhu.shangzhong@zte.com.cn>