gpfs: python2 -> python3, .spec file fix
Fix the overlooked a py2->py3 change in .../scripts/gpfs-epoch

Also the nfs-ganesha.spec file used the IBM package name for a Requires:,
i.e.  gpfs.nfs-ganesha. But I don't believe that just because the -utils
rpm is installed that the -gpfs RPM should also be installed. Rather than
simply fix the name, remove the Requires: completely.

Note that the gpfs-epoch utility is included in the -gpfs RPM.

Also RPM building will automatically derive Requires: python3 from the
BuildRequires: python3-devel, it is not necessary (and even error prone)
to specify the Requires:.

Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>
Change-Id: Idbe7f36a200b408c776f2c20ff179e9169a0f123