python: run setup.py for gpfs-epoch(.py) to fix its shebang
`setup.py build ...` modifies the shebang appropriately for the
platform, i.e. using the python (e.g. /usr/bin/python3) that setup.py
was invoked with.

This might seem like a heavyweight solution for one file, but it's
ostensibly better than some other solution like sed editing the shebang
in an rpmbuild scriptlet. This way you get (almost) the right thing in
a regular cmake/make/make-install build. Almost, because the `setup.py
install ...` step installs gpfs-epoch in /usr/bin instead of where we
infer it's wanted, in /usr/libexec, based on the install scriptlets in
the rpm .spec file.

I contemplated changing the --prefix for the `setup.py install ...` step
but then that scribbles other things in /usr/libexec or /usr/local/libexec

I also contemplated putting gpfs-epoch(.py) in its own subdir, e.g.,
.../src/scripts/gpfs-epoch/... or .../src/FSAL/FSAL_GPFS/gpfs-epoch/...
and could be convinced to do that.

I would prefer to not collapse the %if suse %else if %rhel8 %else %endif
because %suse is actually a lot more complicated. Fixing that is a
different bug for a different patch when I have more $time to untangle
it properly.

Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>
Change-Id: I5ff746393b7bac53c9f12c9d33f0b0aeb24cd5ce