Update patch set 2
Patch Set 2:

> I'm worried about the requirement for 1.1.7.  This means that no stable version of Ganesha will build against any released version of ceph, making FSAL_RGW useless until there is a ceph release and it gets packaged.

worse than that - https://github.com/ceph/ceph/commit/6bf2786d28bad9c5269d07e00142b57b287089f0 changed the API *after* rgw_file version was set to 1.1.7, so the version check isn't quite enough : I have 1.1.7 in that file but it fails building to API conflict.

The ceph side should bump the version to at least 1.1.8 (is an API change really a minor?) and we then need to check for that; but meanwhile the rgw fsal cannot be compiled with any release of ceph as pointed out... I've only pushed two patches and am already annoyed at the CI failures :)
(the CEA CI would also fail on this if reactivated since we build everything buildable to get approved by clang, so we'll need to disable RGW there too...)


Anyway, this patch got merged without +2 because Frank pulled the tip of the branch which had one, I'm not sure it was merged on purpose. If nothing else Dan's comment on the #error version needs fixing.

Patch-set: 2
CC: Gerrit User 1003556 <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>