Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> I don't understand what problem you're trying to solve here. What's ending up blocked when the client doesn't send a RECLAIM_COMPLETE?
> 
> If it's just the buggy client itself, then I don't understand why we should care. If a NFSv4.1+ client does not send a RECLAIM_COMPLETE, then it's broken, period. I don't understand why we'd want to work around such a thing.

An OPEN with CLAIM_NULL will be stuck because we did not receive RECLAIM_COMPLETE. And ganesha will keep returning NFS4ERR_GRACE indefinitely.

I agree that if it is a buggy client then we should fix it. But I feel client not able to complete its reclaim within the grace period when ganesha restarts and the one which does not send RECLAIM_COMPLETE at all are pretty similar.

In case of ganesha restarts, we used to fail the reclaim requests after the grace period ends. I wanted to have similar behaviour so that the client is not stuck indefinitely.

We are using linux kernel client and have seen this issue multiple times, though I am not sure of the exact steps to reproduce the issue.

Patch-set: 2
Reviewer: Gerrit User 1010034 <1010034@d5d70762-12d0-45a1-890d-524b12d3f735>