Update patch set 2
Patch Set 2:

Is it the client that hasn't sent RECLAIM_COMPLETE yet that is trying a non-reclaim open? That's a buggy client.

If one client is still making progress on reclaim and another had completed and is trying to acquire state then NFS4ERR_GRACE is appropriate.

The RECLAIM_COMPLETE operation has the advantage of making it more reasonable to extend grace period a bit when a client has a lot of state to reclaim (note that I'm not sure we're actually doing that part yet...).

Patch-set: 2
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>