Change nfs3_FSALattr_To_Fattr to nfs3_Fixup_FSALattr to save copy
In a future patch, we will serialize READDIRPLUS response as we go
so we would like to save copying the attributes. By moving the copy
out to the caller, we will be able to do that.

In order to save the copy, we add an fsid3 field to struct attrlist
so we can fixup the fsid in the copy passed up on the readdir callback
without damaging the cached attributes.

Change-Id: I61f58c65ff762f0d328cceafd9e99f9aeae44d2b
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>