Make NFSv3 READDIR use xdr_putbuffs
Encode all the directory entries into an xdrmem buffer that can then
be added to the ioq via xdr_putbufs.

We also do a more exact calulation of the space available for the
directory entries so significantly more entries can be returned
if the entries are small.

NOTE: At the moment, this triggers an error in ntirpc because it
tries to add entries to the xdrmem buffer until there is a problem
with them fitting. xdr_string_encode logs an error if putting the
length of the name into the buffer fails.

Change-Id: I456cfd7dd6cf3efd129c9657844878660a140434
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>