Update patch set 1
Patch Set 1:

I'm not sure about this one...  How useful is this?  DNS should be fast, and if it's not, then it can be diagnosed by other means on the same host as Ganesha is running.  This adds a lot of unconditional clock_gettime() calls, which are not cheap, for something that I don't think would be very useful.

When is this useful, and to what extent?  I'm willing to be convinced, if there's a usecase I haven't thought about.

Patch-set: 1
CC: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>