cmake: RFC DO NOT MERGE: FindNTIRPC.cmake
I've been fighting off and on with the FindNTIRPC in the 3.0 RCs
on various distributions when using the unbundled (or sys) libntirpc
packages, but it was hit-or-miss, it worked sometimes, and not others,
so I hacked my way through it and didn't raise any flags about it.

But it came to a head again when building Ubuntu packages on Launchpad,
where it appears that a Debian/Ubuntu-ism is breaking it. I.e. because
unlike Fedora/RHEL/CentOS, which install libraries in /usr/lib64 on
64-bit machines, Debian installs them in /usr/lib/x86_64-linux-gnu/.
(Or /usr/lib/aarch64-linux-gnu/, etc.)

Off hand I don't really understand what the purpose of the first few
if statements is, but mostly it appears that they don't really find
the path to the library, at least on Debian. The use of the PATH_SUFFIXES
seems generally wrong too AFAICT from the find_path and find_library
man pages.

The real issue seems to be in the find_library() invocations, which
having failed to determine the path from find_path() then don't find
the library.

I finally came up with this patch, which worked well enough on Ubuntu
and Debian to get it to build. I haven't tried it on Fedora/RHEL/CentOS
yet.

Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>
Change-Id: I2875ad6048df120fc07b395a2eb88a3ff141424a